# docs/swagger.json
{
    "openapi" : "3.0.1",
    "info" : {
      "title" : "defaultTitle",
      "description" : "defaultDescription",
      "version" : "0.1"
    },
    "servers" : [ {
      "url" : "http://localhost:3000"
    } ],
    "paths" : {
...
      "/products" : {
        "get" : {
          "description" : "Auto generated using Swagger Inspector",
          "parameters" : [ {
            "name" : "category",
            "in" : "query",
            "required" : false,
            "style" : "form",
            "explode" : true,
            "schema" : {
              "type" : "string"
            },
            "example" : "electronics"
          } ],
          "responses" : {
            "200" : {
              "description" : "Auto generated using Swagger Inspector",
              "content" : {
                "application/json; charset=utf-8" : {
                  "schema" : {
                    "type" : "string"
                  },
                  "examples" : { }
                }
              }
            }
          },
          "servers" : [ {
            "url" : "http://localhost:3000"
          } ]
        },
        "post" : {
          "description" : "Auto generated using Swagger Inspector",
          "requestBody" : {
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/body_1"
                },
                "examples" : {
                  "0" : {
                    "value" : "{\n        \"id\": 5,\n        \"category\": \"electronics\",\n        \"name\": \"TV\",\n        \"display_name\": \"OLED TV\",\n        \"description\": \"High Res\"\n    }"
                  }
                }
              }
            }
          },
          "responses" : {
            "201" : {
              "description" : "Auto generated using Swagger Inspector",
              "content" : {
                "application/json; charset=utf-8" : {
                  "schema" : {
                    "type" : "string"
                  },
                  "examples" : { }
                }
              }
            }
          },
          "servers" : [ {
            "url" : "http://localhost:3000"
          } ]
        },
        "servers" : [ {
          "url" : "http://localhost:3000"
        } ]
      },
...
    },
    "components" : {
      "schemas" : {
...
        "body_1" : {
          "type" : "object",
          "properties" : {
            "name" : {
              "type" : "string"
            },
            "description" : {
              "type" : "string"
            },
            "id" : {
              "type" : "integer"
            },
            "category" : {
              "type" : "string"
            },
            "display_name" : {
              "type" : "string"
            }
          }
        },
...
      }
    }
  }